+++ DeadlockCalculator/Models/Character/GunStats.cs
@@ -2,10 +2,10 @@ namespace DeadlockCalculator.Models
 {
     public class GunStats
     {
-        public double BulletDamage { get; set; }
-        public double BulletPerSecond { get; set; }
+        public decimal BulletDamage { get; set; }
+        public decimal BulletPerSecond { get; set; }
         public int Ammo { get; set; }
-        public double ReloadTime { get; set; }
+        public decimal ReloadTime { get; set; }
         public int BulletVelocity { get; set; } 
         public int BulletLifesteal { get; set; }
         public int WeaponDamage { get; set; }
