+++ DeadlockCalculator/Models/Character/SpiritStats.cs
@@ -2,11 +2,11 @@ namespace DeadlockCalculator.Models
 {
     public class SpiritStats
     {
-        public double SpiritPower { get; set; }
-        public double AbilityCooldown { get; set; }
-        public double AbilityRange { get; set; }
+        public decimal SpiritPower { get; set; }
+        public decimal AbilityCooldown { get; set; }
+        public decimal AbilityRange { get; set; }
         public int MaxChargesIncrease { get; set; }
-        public double AbilityDuration { get; set; }
+        public decimal AbilityDuration { get; set; }
         public int SpiritLifesteal { get; set; } 
         public int ChargeCooldown { get; set; }
